BACKPORT: Finish converting to new asm!() syntax.
I got a number of compile failures with `llvm_asm`
with the latest nightly.  Instead of figuring out
how to fix that, I just went ahead and finished
converting everything to use `asm`, which has the
added bonus of being stable, anyway!

I'm not entirely sure this is all correct;
`cargo test` seems to drag in an explicit
dependency on x86-0.44.0, which still fails.

Note also that the SGX modifications are a bit
more elaborate than what had been there; apparently
LLVM uses the %rbx internally, even though the
instructions exposed in that module use that
register.  I worked around that by pushing %rbx
and then copying the leaf operand from %rsi to
%rbx and copying %rbx _back_ to %rsi for output
and popping the original value of %rbx.  The
input/output operands are then writen in terms
of %rsi.

Signed-off-by: Dan Cross <cross@gajendra.net>